RIA-4038 make some changes in code to run test cases
