+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/caseofficer/CaseOfficerRespondentEvidenceSubmittedPersonalisation.java
@@ -1,25 +1,18 @@
 package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.caseofficer;
 
 import static java.util.Objects.requireNonNull;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.APPELLANT_FAMILY_NAME;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.APPELLANT_GIVEN_NAMES;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.LEGAL_REPRESENTATIVE_EMAIL_ADDRESS;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.LEGAL_REP_COMPANY;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.LEGAL_REP_COMPANY_ADDRESS;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.LEGAL_REP_NAME;
-import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.LEGAL_REP_REFERENCE_NUMBER;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.*;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.AddressFormatter.formatCompanyAddress;
 
 import com.google.common.collect.ImmutableMap;
 import java.util.Collections;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import javax.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCase;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.field.AddressUk;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.EmailNotificationPersonalisation;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.service.FeatureToggler;
 import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.EmailAddressFinder;
@@ -82,78 +75,4 @@ public class CaseOfficerRespondentEvidenceSubmittedPersonalisation implements Em
                 .build();
     }
 
-    public String formatCompanyAddress(AsylumCase asylumCase) {
-
-        StringBuilder str = new StringBuilder();
-
-        if (asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class).isPresent()) {
-
-            final String addressLine1 =
-                    asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                            .flatMap(AddressUk::getAddressLine1).orElse("");
-
-            final String addressLine2 =
-                    asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                            .flatMap(AddressUk::getAddressLine2).orElse("");
-
-            final String addressLine3 =
-                    asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                            .flatMap(AddressUk::getAddressLine3).orElse("");
-
-            final String postTown =
-                    asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                            .flatMap(AddressUk::getPostTown).orElse("");
-
-            final String county =
-                    asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                            .flatMap(AddressUk::getCounty).orElse("");
-
-            final String postCode =
-                    asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                            .flatMap(AddressUk::getPostCode).orElse("");
-
-            final String country =
-                    asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                            .flatMap(AddressUk::getCountry).orElse("");
-
-            if (!Optional.of(addressLine1).get().equals("")) {
-                str.append(addressLine1);
-                str.append(", ");
-            }
-
-            if (!Optional.of(addressLine2).get().isEmpty()) {
-                str.append(addressLine2);
-                str.append(", ");
-            }
-
-            if (!Optional.of(addressLine3).get().isEmpty()) {
-                str.append(addressLine3);
-                str.append(", ");
-            }
-
-            if (!Optional.of(postTown).get().isEmpty()) {
-                str.append(postTown);
-                str.append(", ");
-            }
-
-            if (!Optional.of(county).get().isEmpty()) {
-                str.append(county);
-                str.append(", ");
-            }
-
-            if (!Optional.of(postCode).get().isEmpty()) {
-                str.append(postCode);
-                str.append(", ");
-            }
-
-            if (!Optional.of(country).get().isEmpty()) {
-                str.append(country);
-            }
-
-        } else {
-            return "";
-        }
-
-        return str.toString();
-    }
 }
