+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/personalisation/homeoffice/HomeOfficeRemoveRepresentationPersonalisation.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.homeof
 
 import static java.util.Objects.requireNonNull;
 import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition.*;
+import static uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.AddressFormatter.formatCompanyAddress;
 
 import com.google.common.collect.ImmutableMap;
 import java.util.*;
@@ -11,7 +12,6 @@ import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCase;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.AsylumCaseDefinition;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.State;
-import uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities.ccd.field.AddressUk;
 import uk.gov.hmcts.reform.iacasenotificationsapi.domain.personalisation.EmailNotificationPersonalisation;
 import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.AppealService;
 import uk.gov.hmcts.reform.iacasenotificationsapi.infrastructure.CustomerServicesProvider;
@@ -121,78 +121,4 @@ public class HomeOfficeRemoveRepresentationPersonalisation implements EmailNotif
             .build();
     }
 
-    public String formatCompanyAddress(AsylumCase asylumCase) {
-
-        StringBuilder str = new StringBuilder();
-
-        if (asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class).isPresent()) {
-
-            final String addressLine1 =
-                asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                    .flatMap(AddressUk::getAddressLine1).orElse("");
-
-            final String addressLine2 =
-                asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                    .flatMap(AddressUk::getAddressLine2).orElse("");
-
-            final String addressLine3 =
-                asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                    .flatMap(AddressUk::getAddressLine3).orElse("");
-
-            final String postTown =
-                asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                    .flatMap(AddressUk::getPostTown).orElse("");
-
-            final String county =
-                asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                    .flatMap(AddressUk::getCounty).orElse("");
-
-            final String postCode =
-                asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                    .flatMap(AddressUk::getPostCode).orElse("");
-
-            final String country =
-                asylumCase.read(LEGAL_REP_COMPANY_ADDRESS, AddressUk.class)
-                    .flatMap(AddressUk::getCountry).orElse("");
-
-            if (!Optional.of(addressLine1).get().equals("")) {
-                str.append(addressLine1);
-                str.append(", ");
-            }
-
-            if (!Optional.of(addressLine2).get().isEmpty()) {
-                str.append(addressLine2);
-                str.append(", ");
-            }
-
-            if (!Optional.of(addressLine3).get().isEmpty()) {
-                str.append(addressLine3);
-                str.append(", ");
-            }
-
-            if (!Optional.of(postTown).get().isEmpty()) {
-                str.append(postTown);
-                str.append(", ");
-            }
-
-            if (!Optional.of(county).get().isEmpty()) {
-                str.append(county);
-                str.append(", ");
-            }
-
-            if (!Optional.of(postCode).get().isEmpty()) {
-                str.append(postCode);
-                str.append(", ");
-            }
-
-            if (!Optional.of(country).get().isEmpty()) {
-                str.append(country);
-            }
-
-        } else {
-            return "";
-        }
-
-        return str.toString();
-    }
 }
